MEDIUM: fixing unit-tests following the addition of Tests with JarResourceCenter

+++ flexo-ui/src/test/java/org/openflexo/OpenflexoTestCaseWithGUI.java
@@ -81,7 +81,7 @@ public abstract class OpenflexoTestCaseWithGUI extends OpenflexoTestCase {
 			FlexoResourceCenterService RCService = serviceManager.getResourceCenterService();
 			List<FlexoResourceCenter> listRC = RCService.getResourceCenters();
 			for (FlexoResourceCenter rc : listRC) {
-				if (rc instanceof DirectoryResourceCenter) {
+				if (rc instanceof DirectoryResourceCenter && ! rc.getResourceCenterEntry().isSystemEntry()) {
 					File RCDirectory = ((DirectoryResourceCenter) rc).getDirectory();
 					RCDirectory.deleteOnExit();
 				}
@@ -117,7 +117,7 @@ public abstract class OpenflexoTestCaseWithGUI extends OpenflexoTestCase {
 		serviceManager = new TestApplicationContext(generateCompoundTestResourceCenter);
 		for (FlexoResourceCenter rc : serviceManager.getResourceCenterService().getResourceCenters()) {
 			// Select the first directory ResourceCenter
-			if (rc instanceof DirectoryResourceCenter){
+			if (rc instanceof DirectoryResourceCenter && ! rc.getResourceCenterEntry().isSystemEntry()){
 				resourceCenter = (DirectoryResourceCenter) rc;
 				break;
 			}
